+++ class_spec_integration.lua
@@ -129,7 +129,7 @@ function DM.ClassSpec:PushConfigToPlater(forcePush)
     -- Add a static flag to prevent showing the error message multiple times
     if not self.errorMessageShown then
       DM:PrintMessage(
-      "Plater connection not found. Click the 'Install Plater Integration' button in the DotMaster window.")
+        "Plater connection not found. Click the 'Install Plater Integration' button in the DotMaster window.")
       self.errorMessageShown = true
     end
     return